requirements.txt
pandas
sqlalchemy
streamlit
langchain
langchain-community
langchain-experimental
langchain-openai
langchain-google-genai
langchain-ollama
python-dotenv
plotly
openpyxl
tabulate
